requirements.txt
coverage==4.4.2
Django==1.11.7
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-jet==1.0.6
djangorestframework==3.7.3
pkg-resources==0.0.0
psycopg2==2.7.3.2
pytz==2017.3
sqlparse==0.2.4
